+++ crates/sncast/src/main.rs
@@ -1,7 +1,8 @@
+use std::borrow::Borrow;
 use std::num::{NonZeroU8, NonZeroU16};
 use std::str::FromStr;
 
-use crate::starknet_commands::declare::declare;
+use crate::starknet_commands::declare::declare_with_artifacts;
 use crate::starknet_commands::declare_from::{ContractSource, DeclareFrom};
 use crate::starknet_commands::deploy::{DeployArguments, DeployCommonArgs};
 use crate::starknet_commands::get::Get;
@@ -49,7 +50,7 @@ use sncast::{
 use starknet_commands::ledger::{self, Ledger};
 use starknet_commands::verify::Verify;
 use starknet_rust::core::types::ContractClass;
-use starknet_rust::core::types::contract::{AbiEntry, SierraClass};
+use starknet_rust::core::types::contract::{AbiEntry, CompiledClass, SierraClass};
 use starknet_rust::core::utils::get_selector_from_name;
 use starknet_rust::providers::Provider;
 use starknet_types_core::felt::Felt;
@@ -240,24 +241,34 @@ pub struct Arguments {
     pub arguments: Option<String>,
 }
 
+enum AbiOrContractClass<C> {
+    Abi(Vec<AbiEntry>),
+    ContractClass(C),
+}
+
 impl Arguments {
     fn try_into_calldata(
         self,
-        contract_class: &ContractClass,
+        abi_or_contract_class: AbiOrContractClass<impl Borrow<ContractClass>>,
         selector: &Felt,
     ) -> Result<Vec<Felt>> {
         if let Some(calldata) = self.calldata {
-            calldata_to_felts(&calldata)
-        } else {
-            let ContractClass::Sierra(sierra_class) = contract_class else {
-                bail!("Transformation of arguments is not available for Cairo Zero contracts")
-            };
-
-            let abi: Vec<AbiEntry> = serde_json::from_str(sierra_class.abi.as_str())
-                .context("Couldn't deserialize ABI received from network")?;
-
-            transform(&self.arguments.unwrap_or_default(), &abi, selector)
+            return calldata_to_felts(&calldata);
         }
+
+        let abi = match abi_or_contract_class {
+            AbiOrContractClass::Abi(abi) => abi,
+            AbiOrContractClass::ContractClass(contract_class) => {
+                let ContractClass::Sierra(sierra_class) = contract_class.borrow() else {
+                    bail!("Transformation of arguments is not available for Cairo Zero contracts")
+                };
+
+                serde_json::from_str::<Vec<_>>(sierra_class.abi.as_str())
+                    .context("Couldn't deserialize ABI received from network")?
+            }
+        };
+
+        transform(&self.arguments.unwrap_or_default(), &abi, selector)
     }
 }
 
@@ -363,6 +374,7 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<Exit
                     declare.common.fee_args,
                     declare.common.dry_run_args,
                     declare.common.nonce,
+                    declare.no_abi,
                     account,
                     &artifacts,
                     wait_config,
@@ -399,6 +411,7 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<Exit
                 let network_flag = generate_network_flag(&rpc, &config);
                 Some(DeployCommandMessage::new(
                     &contract_definition.abi,
+                    declare.no_abi,
                     response,
                     &config.account,
                     &config.accounts_file,
@@ -441,6 +454,7 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<Exit
             let result = with_account!(&account, |account| {
                 starknet_commands::declare_from::declare_from(
                     contract_source,
+                    declare_from.no_abi,
                     &declare_from.common,
                     account,
                     wait_config,
@@ -474,6 +488,10 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<Exit
         }
 
         Commands::Deploy(deploy) => {
+            if deploy.common.contract_identifier.contract_name.is_none() && deploy.no_abi {
+                bail!("`--no-abi` can only be used with `--contract-name`");
+            }
+
             let Deploy {
                 common:
                     DeployCommonArgs {
@@ -486,6 +504,7 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<Exit
                 dry_run_args,
                 rpc,
                 mut nonce,
+                no_abi,
                 ..
             } = deploy;
 
@@ -493,8 +512,10 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<Exit
 
             let account = get_account(&config, &provider, &rpc, ui).await?;
 
-            let (class_hash, declare_response) = if let Some(class_hash) = identifier.class_hash {
-                (class_hash, None)
+            let (class_hash, declare_response, local_abi) = if let Some(class_hash) =
+                identifier.class_hash
+            {
+                (class_hash, None, None)
             } else if let Some(contract_name) = identifier.contract_name {
                 let manifest_path = assert_manifest_path_exists()?;
                 let package_metadata = get_package_metadata(&manifest_path, &package)?;
@@ -511,14 +532,26 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<Exit
                 )
                 .expect("Failed to build contract");
 
+                let contract_artifacts = artifacts
+                    .get(&contract_name)
+                    .expect("Failed to get contract artifacts");
+                let contract_definition: SierraClass =
+                    serde_json::from_str(&contract_artifacts.sierra)
+                        .context("Failed to parse sierra artifact")?;
+                let casm_contract_definition: CompiledClass =
+                    serde_json::from_str(&contract_artifacts.casm)
+                        .context("Failed to parse casm artifact")?;
+                let local_abi = contract_definition.abi.clone();
+
                 let declare_result = with_account!(&account, |account| {
-                    declare(
-                        contract_name,
-                        fee_args,
-                        dry_run_args,
+                    declare_with_artifacts(
+                        contract_definition,
+                        casm_contract_definition,
+                        &fee_args,
+                        &dry_run_args,
                         nonce,
+                        no_abi,
                         account,
-                        &artifacts,
                         WaitForTx {
                             wait: true,
                             wait_params: wait_config.wait_params,
@@ -537,10 +570,11 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<Exit
                 match declare_result {
                     Ok(DeclareResponse::AlreadyDeclared(AlreadyDeclaredResponse {
                         class_hash,
-                    })) => (class_hash.into_(), None),
+                    })) => (class_hash.into_(), None, Some(local_abi)),
                     Ok(DeclareResponse::Success(declare_transaction_response)) => (
                         declare_transaction_response.class_hash.into_(),
                         Some(declare_transaction_response),
+                        Some(local_abi),
                     ),
                     Ok(DeclareResponse::DryRun(_)) => {
                         unreachable!(
@@ -565,10 +599,16 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<Exit
             // safe to unwrap because "constructor" is a standardized name
             let selector = get_selector_from_name("constructor").unwrap();
 
-            let contract_class = get_contract_class(class_hash, &provider).await?;
-
             let arguments: Arguments = arguments.into();
-            let calldata = arguments.try_into_calldata(&contract_class, &selector)?;
+            let calldata = arguments.try_into_calldata(
+                match local_abi {
+                    Some(abi) => AbiOrContractClass::Abi(abi),
+                    None => AbiOrContractClass::ContractClass(
+                        get_contract_class(class_hash, &provider).await?,
+                    ),
+                },
+                &selector,
+            )?;
 
             let result = with_account!(&account, |account| {
                 starknet_commands::deploy::deploy(
@@ -624,7 +664,10 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<Exit
             let selector = get_selector_from_name(&function)
                 .context("Failed to convert entry point selector to FieldElement")?;
 
-            let calldata = arguments.try_into_calldata(&contract_class, &selector)?;
+            let calldata = arguments.try_into_calldata(
+                AbiOrContractClass::ContractClass(&contract_class),
+                &selector,
+            )?;
 
             let result = starknet_commands::call::call(
                 contract_address,
@@ -677,7 +720,10 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<Exit
             let class_hash = get_class_hash_by_address(&provider, contract_address).await?;
             let contract_class = get_contract_class(class_hash, &provider).await?;
 
-            let calldata = arguments.try_into_calldata(&contract_class, &selector)?;
+            let calldata = arguments.try_into_calldata(
+                AbiOrContractClass::ContractClass(&contract_class),
+                &selector,
+            )?;
 
             let result = with_account!(&account, |account| {
                 starknet_commands::invoke::invoke(